# Gzip TRAP files