@jhutchings1 2020-03-11
Thanks for doing this @max-schaefer ! ✨ This will definitely help to ensure more analyses can be completed on the hosted Actions runner. https://help.github.com/en/actions/reference/virtual-environments-for-github-hosted-runners