@sauyon 2020-03-11
Is this something we might want an option for in the future?