@max-schaefer 2020-03-12
> Is this something we might want an option for in the future?

No. All other extractors zip their trap files, we should do so, too.