@max-schaefer 2020-03-12
Yes, I've set off a dist-compare just to make sure (I don't trust myself around streams, I'm afraid).